Move Planner stuff to Core
This is prep for any structural changes needed for finishing swipe stuff.

refs: MBL-14016
affects: Parent
release note: none
