+++ Core/Core/Planner/CalendarDaysViewController.swift
@@ -18,7 +18,6 @@
 
 import Foundation
 import UIKit
-import Core
 
 protocol CalendarDaysDelegate: class {
     func setSelectedDate(_ date: Date)
