+++ Core/CoreTests/Planner/APIPlannableTests.swift
@@ -1,6 +1,6 @@
 //
 // This file is part of Canvas.
-// Copyright (C) 2018-present  Instructure, Inc.
+// Copyright (C) 2020-present  Instructure, Inc.
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Affero General Public License as
@@ -19,7 +19,7 @@
 import XCTest
 @testable import Core
 
-class GetPlannablesRequestTests: XCTestCase {
+class APIPlannableTests: XCTestCase {
     var req: GetPlannablesRequest!
 
     override func setUp() {
