+++ Parent/Parent/Dashboard/DashboardViewController.swift
@@ -311,7 +311,7 @@ class DashboardViewController: UIViewController, CustomNavbarProtocol {
     func calendarViewController(_ session: Session, startDate: Date = Date()) -> UIViewController? {
         guard let currentStudent = currentStudent else { return nil }
         if ExperimentalFeature.parentCalendar.isEnabled {
-            return CalendarContainerViewController.create(studentID: currentStudent.id)
+            return PlannerViewController.create(studentID: currentStudent.id)
         }
         return CalendarEventWeekPageViewController.create(session: session, studentID: currentStudent.id, initialReferenceDate: startDate)
     }
