+++ Parent/ParentUnitTests/RoutesTests.swift
@@ -28,7 +28,7 @@ class RoutesTests: ParentTestCase {
         XCTAssert(Parent.router.match(Route.conversation("1").url) is ConversationDetailViewController)
 
         ExperimentalFeature.parentCalendar.isEnabled = true
-        XCTAssert(Parent.router.match(.parse("/calendar")) is CalendarContainerViewController)
+        XCTAssert(Parent.router.match(.parse("/calendar")) is PlannerViewController)
         ExperimentalFeature.parentCalendar.isEnabled = false
         XCTAssertNil(Parent.router.match(.parse("/calendar")))
     }
